requirements.txt
# Broker APIs
# fyers-apiv3 hard-pins aiohttp==3.9.3, conflicting with growwapi>=3.11.18
# Install via: pip install fyers-apiv3 --no-deps (handled in install.sh)
growwapi>=1.0.0
pyotp>=2.9.0

# Data
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=15.0.0
fastparquet>=2023.10.0
yfinance>=0.2.36

# ML / AI
scikit-learn>=1.4.0
xgboost>=2.0.0
tensorflow>=2.15.0
joblib>=1.3.0
shap>=0.44.0
TA-Lib>=0.4.28

# Options / Greeks
mibian>=0.1.3

# Web Framework
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
aiohttp>=3.11.18

# Database
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# Config / Validation
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# Timezone
pytz>=2024.1

# Alerts
python-telegram-bot>=20.0

# Utils
requests>=2.31.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.13.0
